move DeleteStatement to pkg/ast
Moves the DeleteStatement struct to `pkg/ast`
