+++ go.mod
@@ -3,8 +3,6 @@ module github.com/jaypipes/sqlb
 go 1.15
 
 require (
-	github.com/go-sql-driver/mysql v1.6.0
-	github.com/hashicorp/hcl v1.0.0
-	github.com/lib/pq v1.10.2
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/stretchr/testify v1.7.0
 )
